setup vitest config and gh flow
