+++ .github/workflows/publish.yml
@@ -28,8 +28,8 @@ jobs:
 
       - name: Install dependencies
         run: pnpm install
-      # - name: Run tests
-        # run: npm test
+      - name: Run tests
+        run: pnpm test
 
       - name: Build package
         run: pnpm build
