+++ src/try-catch.test.ts
@@ -1,0 +1,10 @@
+import { describe, test } from 'vitest';
+
+describe('try-catch', () => {
+    test.skip('should have successfully retrieved data from fetch', () => {});
+    test.skip('should have error not be null from fetch', () => {});
+
+    test.skip('should have successfully resolved promise', () => {});
+    test.skip('should have error not be null from rejected promise', () => {});
+});
+
